+++ common/src/main/java/tocraft/walkers/mixin/player/PlayerEntityDataMixin.java
@@ -49,6 +49,9 @@ public abstract class PlayerEntityDataMixin extends LivingEntity implements Play
     @Nullable
     private LivingEntity walkers$shape = null;
 
+    @Unique
+    private float walkers$normalHealth = -1;
+
     private PlayerEntityDataMixin(EntityType<? extends LivingEntity> type, Level world) {
         super(type, world);
     }
@@ -65,6 +68,9 @@ public abstract class PlayerEntityDataMixin extends LivingEntity implements Play
 
         // Hostility
         walkers$remainingTime = tag.getInt("RemainingHostilityTime");
+
+        // Health
+        walkers$normalHealth = tag.getFloat("walkersHealth");
 
         // Current Walkers
         walkers$readCurrentShape(tag.getCompound("CurrentShape"));
@@ -83,6 +89,9 @@ public abstract class PlayerEntityDataMixin extends LivingEntity implements Play
 
         // Hostility
         tag.putInt("RemainingHostilityTime", walkers$remainingTime);
+
+        // Health
+        tag.putFloat("walkersHealth", walkers$normalHealth);
 
         // Current Walkers
         tag.put("CurrentShape", walkers$writeCurrentShape(new CompoundTag()));
@@ -200,6 +209,8 @@ public abstract class PlayerEntityDataMixin extends LivingEntity implements Play
         // health to reflect shape.
         if (shape != null) {
             if (Walkers.CONFIG.scalingHealth && healthAttribute != null) {
+                this.walkers$normalHealth = player.getMaxHealth();
+
                 // calculate the current health in percentage, used later
                 float currentHealthPercent = player.getHealth() / player.getMaxHealth();
 
@@ -226,10 +237,26 @@ public abstract class PlayerEntityDataMixin extends LivingEntity implements Play
         // If the shape is null (going back to player), set the player's base health
         // value to 20 (default) to clear old changes.
         if (shape == null) {
-            float currentHealthPercent = player.getHealth() / player.getMaxHealth();
-
             if (Walkers.CONFIG.scalingHealth && healthAttribute != null) {
-                healthAttribute.setBaseValue(20);
+                float currentHealthPercent = player.getHealth() / player.getMaxHealth();
+
+                float health = Walkers.CONFIG.player_health;
+                if (health < 0) {
+                    if (walkers$normalHealth > 0) {
+                        health = walkers$normalHealth;
+                    } else {
+                        health = 20;
+                    }
+                }
+                healthAttribute.setBaseValue(health);
+
+                // Clear health value if needed
+                if (Walkers.CONFIG.percentScalingHealth) {
+
+                    player.setHealth(Math.min(currentHealthPercent * player.getMaxHealth(), player.getMaxHealth()));
+                } else {
+                    player.setHealth(Math.min(player.getHealth(), player.getMaxHealth()));
+                }
             }
             if (Walkers.CONFIG.scalingAmor) {
                 if (armorAttribute != null) {
@@ -238,14 +265,6 @@ public abstract class PlayerEntityDataMixin extends LivingEntity implements Play
                 if (armorToughnessAttribute != null) {
                     armorAttribute.setBaseValue(0);
                 }
-            }
-
-            // Clear health value if needed
-            if (Walkers.CONFIG.percentScalingHealth) {
-                player.setHealth(Math.min(currentHealthPercent * player.getMaxHealth(), player.getMaxHealth()));
-            }
-            else {
-                player.setHealth(Math.min(player.getHealth(), player.getMaxHealth()));
             }
         }