+++ CHANGELOG.md
@@ -3,6 +3,7 @@ walkers 5.8.0
 - add AlexsCaves mobs to flying.json
 - add a config option to prevent hostile mobs from attacking neutral
 - fix showPlayerNametag with Team settings
+- add a config option to manually adapt the player health / cache it
 
 walkers 5.7
 ================